int8 to int

+++ include/phy/mapper.h
@@ -9,6 +9,6 @@
 using complex_d = std::complex<double>; 
 using complex_int16 = std::complex<int16_t>; 
 
-std::vector<complex_d> bpsk(const std::vector <int8_t> &bits);
+std::vector<complex_d> bpsk(const std::vector <int> &bits);
 
-std::vector<complex_d> qpsk(const std::vector <int8_t> &bits);
+std::vector<complex_d> qpsk(const std::vector <int> &bits);
